pyproject.toml
[tool.poetry]
name = "flagsmith"
version = "4.0.1"
description = "Flagsmith Python SDK"
authors = ["Flagsmith <support@flagsmith.com>"]
license = "BSD3"
readme = "README.md"
keywords = ["feature", "flag", "flagsmith", "remote", "config"]
documentation = "https://docs.flagsmith.com"
packages = [{ include = "flagsmith" }]

[tool.poetry.dependencies]
flagsmith-flag-engine = { git = "https://github.com/Flagsmith/flagsmith-engine.git", branch = "feat/generic-metadata" }
python = ">=3.9,<4"
requests = "^2.32.3"
requests-futures = "^1.0.1"
sseclient-py = "^1.8.0"
typing-extensions = "^4.15.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
mypy = { version = "^1.16.1", python = ">=3.9,<4" }
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.6.1"
pre-commit = { version = "^4.2.0", python = ">=3.9,<4" }
responses = "^0.24.1"
types-requests = "^2.32"
pyfakefs = "^5.9.2"

[tool.mypy]
exclude = ["example/*"]

[tool.black]
target-version = ["py39"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
